requirements.txt
# YouTube Short Generator – dependencies
# Install: pip install -r requirements.txt

pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
Pillow>=10.0.0
moviepy>=1.0.3
edge-tts>=6.1.0
